# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 209jobs API
  description: |
    # 209jobs API Documentation

    Welcome to the 209jobs API! This RESTful API provides comprehensive job board functionality including:

    - **Job Management**: Create, update, search, and manage job postings
    - **User Management**: Handle user profiles, preferences, and authentication
    - **Alert System**: Smart job alerts with AI-powered matching
    - **Advertisement Platform**: Comprehensive ad management with tracking
    - **Application Tracking**: Manage job applications and candidate interactions

    ## Authentication

    The API uses JWT-based authentication with role-based access control:
    - **Jobseekers**: Can search jobs, apply, create alerts
    - **Employers**: Can post jobs, manage applications, create ads
    - **Admins**: Full system access

    ## Rate Limiting

    API endpoints are rate-limited to ensure fair usage:
    - **Standard endpoints**: 100 requests per minute
    - **Search endpoints**: 500 requests per minute  
    - **Tracking endpoints**: 1000 requests per minute
    - **Auth endpoints**: 200 requests per minute

    ## Error Handling

    All endpoints return consistent error responses with appropriate HTTP status codes and detailed error messages.

  version: 1.0.0
  contact:
    name: 209jobs API Support
    url: https://209jobs.com/support
    email: api-support@209jobs.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: Search jobs
      description: |
        Search and filter job listings with advanced options including:
        - Full-text search across title and description
        - Location-based filtering with radius support
        - Salary range filtering
        - Job type and experience level filters
        - Skills and company filtering
        - Sorting and pagination
      security: []
      parameters:
        - $ref: '#/components/parameters/QueryParam'
        - $ref: '#/components/parameters/LocationParam'
        - $ref: '#/components/parameters/JobTypeParam'
        - $ref: '#/components/parameters/ExperienceLevelParam'
        - $ref: '#/components/parameters/SalaryMinParam'
        - $ref: '#/components/parameters/SalaryMaxParam'
        - $ref: '#/components/parameters/RemoteParam'
        - $ref: '#/components/parameters/SkillsParam'
        - $ref: '#/components/parameters/CompanyParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Jobs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    post:
      tags:
        - Jobs
      summary: Create job posting
      description: Create a new job posting (employers only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /jobs/semantic-search:
    post:
      tags:
        - Jobs
      summary: Semantic job search
      description: Advanced AI-powered job search with semantic matching
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSearchRequest'
      responses:
        '200':
          description: Semantic search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # User Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  parameters:
...
    QueryParam:
      name: query
      in: query
      description: Search query
      schema:
        type: string
        maxLength: 200
    LocationParam:
      name: location
      in: query
      description: Location filter
      schema:
        type: string
        maxLength: 100
    JobTypeParam:
      name: jobType
      in: query
      description: Job type filter
      schema:
        $ref: '#/components/schemas/JobType'
    ExperienceLevelParam:
      name: experienceLevel
      in: query
      description: Experience level filter
      schema:
        $ref: '#/components/schemas/ExperienceLevel'
    SalaryMinParam:
      name: salaryMin
      in: query
      description: Minimum salary filter
      schema:
        type: integer
        minimum: 0
    SalaryMaxParam:
      name: salaryMax
      in: query
      description: Maximum salary filter
      schema:
        type: integer
        minimum: 0
    RemoteParam:
      name: remote
      in: query
      description: Remote work filter
      schema:
        type: boolean
    SkillsParam:
      name: skills
      in: query
      description: Skills filter (comma-separated)
      schema:
        type: string
    CompanyParam:
      name: company
      in: query
      description: Company filter
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [created_desc, created_asc, salary_desc, salary_asc, relevance]
        default: created_desc
  schemas:
...
    CreateJobRequest:
      type: object
      required:
        - title
        - company
        - description
        - requirements
        - location
        - jobType
        - experienceLevel
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        company:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 50
          maxLength: 10000
        requirements:
          type: string
          minLength: 1
          maxLength: 5000
        location:
          type: string
          minLength: 1
          maxLength: 200
        jobType:
          $ref: '#/components/schemas/JobType'
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        salaryMin:
          type: integer
          minimum: 0
        salaryMax:
          type: integer
          minimum: 0
        benefits:
          type: string
          maxLength: 2000
        skills:
          type: array
          items:
            type: string
          maxItems: 50
        isRemote:
          type: boolean
          default: false
        applicationUrl:
          type: string
          format: uri
        applicationEmail:
          type: string
          format: email
        expiresAt:
          type: string
          format: date-time
...
    SemanticSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Natural language job search query
        location:
          type: string
          maxLength: 200
        radius:
          type: integer
          minimum: 1
          maximum: 100
          description: Search radius in miles
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        company:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        benefits:
          type: string
        skills:
          type: array
          items:
            type: string
        isRemote:
          type: boolean
        applicationUrl:
          type: string
          format: uri
        applicationEmail:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/JobStatus'
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        employerId:
          type: string
          format: uuid
        applicationCount:
          type: integer
          description: Number of applications received
    JobListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            jobs:
              type: array
              items:
                $ref: '#/components/schemas/Job'
            pagination:
              $ref: '#/components/schemas/Pagination'
            filters:
              type: object
              description: Available filter options
            totalCount:
              type: integer
...
    JobResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Job'
    SemanticSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            jobs:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Job'
                  - type: object
                    properties:
                      relevanceScore:
                        type: number
                        minimum: 0
                        maximum: 1
                        description: AI-calculated relevance score
                      matchedFields:
                        type: array
                        items:
                          type: string
                        description: Fields that matched the query
            query:
              type: string
              description: Processed query
            totalResults:
              type: integer

    # User Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time

    # Enum Schemas
...
    JobType:
      type: string
      enum: [full-time, part-time, contract, temporary, internship]
    ExperienceLevel:
      type: string
      enum: [entry, mid, senior, executive]
    JobStatus:
      type: string
      enum: [active, paused, completed, expired]
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Validation failed'
            errors:
              email: ['Invalid email format']
              password: ['Password must be at least 8 characters']
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Authentication required'
            code: 'UNAUTHORIZED'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Insufficient permissions'
            code: 'FORBIDDEN'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Rate limit exceeded. Try again later.'
            code: 'RATE_LIMIT_EXCEEDED'
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
...